# src/za.co.grindrodbank.a3s/wwwroot/openapi-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "description" : "API Definition for the A3S. This service allows authentication, authorisation and accounting.",
    "title" : "A3S",
    "version" : "1.0.5"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/clients" : {
      "get" : {
        "description" : "Search for clients.",
        "operationId" : "listClients",
        "parameters" : [ {
          "description" : "The page to view.",
          "explode" : true,
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "description" : "The size of a page.",
          "explode" : true,
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 10,
            "format" : "int32",
            "maximum" : 20,
            "minimum" : 1,
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "description" : "A search query filter on the name",
          "example" : "FilterByEntityType",
          "explode" : true,
          "in" : "query",
          "name" : "filterName",
          "required" : false,
          "schema" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "style" : "form"
        }, {
          "description" : "a comma separated list of fields in their sort order. Ascending order is assumed. Append desc after a field to indicate descending order.",
          "explode" : false,
          "in" : "query",
          "name" : "orderBy",
          "required" : false,
          "schema" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Oauth2Client"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Pagination" : {
                "description" : "Pagination header.",
                "explode" : false,
                "schema" : {
                  "$ref" : "#/components/schemas/PaginationHeaderResponse"
                },
                "style" : "simple"
              }
            }
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Bad Request."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not authenticated."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden - You are not authorized to access the list of clients."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Client list not found."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "An unexpected error occurred."
          }
        },
        "summary" : "Search for clients.",
        "tags" : [ "Client" ]
      }
    },
    "/clients/{clientId}" : {
      "get" : {
        "description" : "Get a client by its client Id.",
        "operationId" : "getClient",
        "parameters" : [ {
          "description" : "client",
          "explode" : false,
          "in" : "path",
          "name" : "clientId",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Oauth2Client"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Bad Request."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not authenticated."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden - You are not authorized to access the client."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Client not found."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "An unexpected error occurred"
          }
        },
        "summary" : "Get a client.",
        "tags" : [ "Client" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Oauth2Client" : {
        "description" : "Models an Oauth2 or OpenID Connect client.\n",
        "example" : {
          "allowedOfflineAccess" : true,
          "allowedGrantTypes" : [ "allowedGrantTypes", "allowedGrantTypes" ],
          "clientId" : "clientId",
          "allowedCorsOrigins" : [ "http://example.com/aeiou", "http://example.com/aeiou" ],
          "postLogoutRedirectUris" : [ "http://example.com/aeiou", "http://example.com/aeiou" ],
          "name" : "name",
          "redirectUris" : [ "http://example.com/aeiou", "http://example.com/aeiou" ],
          "allowedScopes" : [ "allowedScopes", "allowedScopes" ]
        },
        "properties" : {
          "clientId" : {
            "description" : "The unique ID of the client.",
            "type" : "string"
          },
          "name" : {
            "description" : "Client display name. Used for logging and the consent screen.",
            "type" : "string"
          },
          "allowedGrantTypes" : {
            "description" : "The Oauth2 grant types that the client will be permitted to use.",
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "redirectUris" : {
            "description" : "Specifies the allowed URIs to return tokens or authorisation codes to.",
            "items" : {
              "format" : "url",
              "type" : "string"
            },
            "type" : "array"
          },
          "allowedCorsOrigins" : {
            "description" : "Sets the allowed CORS origins for JavaScript clients.",
            "items" : {
              "format" : "url",
              "type" : "string"
            },
            "type" : "array"
          },
          "postLogoutRedirectUris" : {
            "description" : "Specifies the allowed URIs to redirect to after logout.",
            "items" : {
              "format" : "url",
              "type" : "string"
            },
            "type" : "array"
          },
          "allowedScopes" : {
            "description" : "Specifies the scopes that the client is allowed to access. If empty, the client cannot access any scopes.",
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "allowedOfflineAccess" : {
            "description" : "Defines whether offline access with refresh tokens is permitted for this client.",
            "type" : "boolean"
          }
        },
        "required" : [ "allowedGrantTypes", "allowedOfflineAccess", "allowedScopes", "clientId", "name" ],
        "type" : "object"
      },
...
      "ErrorResponse" : {
        "description" : "Error response object to give more info with HTTP status code",
        "properties" : {
          "message" : {
            "type" : "string"
          }
        },
        "required" : [ "message" ],
        "type" : "object"
      },
      "PaginationHeaderResponse" : {
        "description" : "This is the standard response for requests that have paginated collections",
        "properties" : {
          "total" : {
            "description" : "The total number of pages in the result set",
            "format" : "int32",
            "minimum" : 1,
            "type" : "integer"
          },
          "count" : {
            "description" : "The total number of results in the result set",
            "format" : "int32",
            "minimum" : 1,
            "type" : "integer"
          },
          "size" : {
            "default" : 10,
            "description" : "The number items to include in a page of results. The page size for the result set",
            "format" : "int32",
            "maximum" : 20,
            "minimum" : 1,
            "type" : "integer"
          },
          "current" : {
            "description" : "The position of the page in the paged result set that is being returned",
            "format" : "int32",
            "minimum" : 1,
            "type" : "integer"
          },
          "first" : {
            "description" : "The link to the first page of results containing [size] results. This link includes the page, size, filter, orderBy and fields query parameters.",
            "format" : "url",
            "type" : "string"
          },
          "last" : {
            "description" : "The link to the last page of results containing [size] results. This link includes the page, size, filter, orderBy and fields query parameters. If this is the first page in the result set then this will be [null].",
            "format" : "url",
            "type" : "string"
          },
          "prev" : {
            "description" : "The link to the previous page of results containing [size] results. This link includes the page, size, filter, orderBy and fields query parameters. If this is the first page in the result set then this will be [null].",
            "format" : "url",
            "type" : "string"
          },
          "next" : {
            "description" : "The link to the next page of results containing [size] results. This link includes the page, size, filter, orderBy and fields query parameters. If this is the last page in the result set then this will be [null].",
            "format" : "url",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  }
}